+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Native MQTT Gateway Docker build
+- MQTT
+  - automatically register for all alarm types
+  - create Docker image
 
 ## [0.0.3] - 2022-12-19
 ### Changed